+++ packages/radio/src/test/mwc-radio.test.ts
@@ -139,5 +139,78 @@ suite('mwc-radio', () => {
       assert.isFalse(a2.checked);
       assert.isFalse(b1.checked);
     });
+
+    test('when checked before connected', () => {
+      const r1 = document.createElement('mwc-radio');
+      r1.name = 'a';
+      const r2 = document.createElement('mwc-radio');
+      r2.name = 'a';
+      const r3 = document.createElement('mwc-radio');
+      r3.name = 'a';
+
+      // r1 and r2 should both be checked, because even though they have the
+      // same name, they aren't yet connected to a root. Groups are scoped to
+      // roots, and we can't know which root a radio belongs to until it is
+      // connected to one. This matches native <input type="radio"> behavior.
+      r1.checked = true;
+      r2.checked = true;
+      assert.isTrue(r1.checked);
+      assert.isTrue(r2.checked);
+      assert.isFalse(r3.checked);
+
+      // Connecting r1 shouldn't change anything, since it's the only one in the
+      // group.
+      container.appendChild(r1);
+      assert.isTrue(r1.checked);
+      assert.isTrue(r2.checked);
+      assert.isFalse(r3.checked);
+
+      // Appending r2 should disable r1, because when a new checked radio is
+      // connected, it wins (this matches native input behavior).
+      container.appendChild(r2);
+      assert.isFalse(r1.checked);
+      assert.isTrue(r2.checked);
+      assert.isFalse(r3.checked);
+
+      // Appending r3 shouldn't change anything, because it's not checked.
+      container.appendChild(r3);
+      assert.isFalse(r1.checked);
+      assert.isTrue(r2.checked);
+      assert.isFalse(r3.checked);
+
+      // Checking r3 should uncheck r2 because it's now in the same group.
+      r3.checked = true;
+      assert.isFalse(r1.checked);
+      assert.isFalse(r2.checked);
+      assert.isTrue(r3.checked);
+    });
+
+    test('in a lit repeat', () => {
+      const values = ['a1', 'a2'];
+      render(
+          html`${
+              values.map(
+                  (value) =>
+                      html`<mwc-radio value=${value} name="a"></mwc-radio>`)}`,
+          container);
+      const [a1, a2] = container.querySelectorAll('mwc-radio');
+
+      assert.isFalse(a1.checked);
+      assert.isFalse(a2.checked);
+      assert.equal(a1.value, values[0]);
+      assert.equal(a2.value, values[1]);
+
+      a1.checked = true;
+      assert.isTrue(a1.checked);
+      assert.isFalse(a2.checked);
+
+      a2.checked = true;
+      assert.isFalse(a1.checked);
+      assert.isTrue(a2.checked);
+
+      a2.checked = false;
+      assert.isFalse(a1.checked);
+      assert.isFalse(a2.checked);
+    });
   });
 });